Fixed a compiler warning in MacOS VideoComponent

+++ modules/juce_video/native/juce_mac_Video.h
@@ -123,7 +123,7 @@ struct VideoComponent::Pimpl   : public BaseClass
 
     void setSpeed (double newSpeed)
     {
-        [getAVPlayer() setRate: newSpeed];
+        [getAVPlayer() setRate: (float) newSpeed];
     }
 
     double getSpeed() const
